+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Version 0.11 (unreleased)
 
 **Major enhancements**
 
+* Optionally output to a user-specified array (``out`` keyword argument) when constructing
+  geometries from ``indices`` (#380).
 * Added ``pygeos.empty`` to create a geometry array pre-filled with None or
   with empty geometries (#381).